mypageに関する処理をmypage.goに集約
